Corrections du build Linux
Corrections de warnings
Ajout des macros pour détecter le Debug vs Release
Changement de `div()` a `std::lldiv()`
Le test data du jour 6 est maintenant dans le .h

+++ day03.cpp
@@ -80,7 +80,7 @@ static Point GetDirection(char action)
 static void Run(std::string_view action, unsigned int mask, Point &pos, int &length, Grid &grid)
 {
     Point dir = GetDirection(action.front());
-    int dist = std::strtol(action.data() + 1, nullptr, 10);
+    int dist = static_cast<int>(std::strtol(action.data() + 1, nullptr, 10));
 
     for (int i = 0; i != dist; ++i)
     {
